fix: css theme for PrimeVue
