+++ frontend/src/components/Map/MapPopup.vue
@@ -1,23 +1,8 @@
 <script setup lang="ts">
 import { useProductStore } from '@/stores/productStore'
-import { computed, ref, onMounted, onBeforeUnmount } from 'vue'
-import { useUITheme } from '@/composables/useUITheme'
+import { computed, ref } from 'vue'
 
 const productStore = useProductStore()
-const { widgetClasses, widgetStyles, initSystemThemeListener } = useUITheme()
-
-// Initialize system theme listener and store cleanup function
-let cleanupThemeListener: (() => void) | null = null
-
-onMounted(() => {
-  cleanupThemeListener = initSystemThemeListener()
-})
-
-onBeforeUnmount(() => {
-  if (cleanupThemeListener) {
-    cleanupThemeListener()
-  }
-})
 
 /**
  * Computed property for the absolute X position of the popup.
@@ -78,14 +63,12 @@ function hidePopup(): void {
     v-if="productStore.clickedPoint && productStore.clickedPoint.show"
     :showCloseIcon="true"
     @hide="hidePopup"
-    :class="['map-popup', widgetClasses]"
-    :style="{ left: absX, top: absY, ...widgetStyles }"
+    :class="['map-popup']"
+    :style="{ left: absX, top: absY }"
     :dismissable="true"
-    :modal="false"
     :autoZIndex="true"
     :baseZIndex="1001"
     appendTo="self"
-    :visible="true"
   >
     <template #content>
       <div class="popup-content">
